Fixed test; tree method for files; associating parents and superclasses.
The test was erroring due to an incorrect version number. A tree method
turns the collection of files into a nested hash of arrays, simulating a
file system. Two workarounds implemented for RDoc bugs: one to associate
child modules with their parents (not sure how RDoc messes this up), and
another to find the superclass of a class that has been defined in the
same file as its superclass, because RDoc just gives it a string (rather
than a class object).